refactor: 드랍다운 리팩토링
